Add new user profile view and increase security in registration
This commit adds a new 'UsersComponent' which introduces a profile view for users. The 'UsersService' is also added to provide various functionalities for the user such as getting user by ID, getting following users, updating posts and getting user's posts.

In the User model, the 'followers' field is changed from unknown type to array of 'Followers' type.

In the registration form found in 'RegisterComponent', validators are also added for the password to have minimum of 8 characters and the email to follow standard email format.

Lastly, a 'reset' function is added to enable clearing of form upon user request.

All the modifications are done to improve user experience, and security during user registration, and to further develop the user profile part of the application.