Update and format post and comment components
PostId is now subscribable in the "UsersPostsComponent", and the "UsersCreateCommentsComponent" has been formatted to implement OnDestroy for better subscription management. The "UsersProfile" model has been removed due to redundancy, with the relevant data being transferred to the general "User" model instead. As a result, the application's performance was improved by reducing unnecessary code and streamlining database requests.