Add a feature to enhance user interfaces and refactor code
Added an "add likes" feature to enhance post interactions from users. Cleaned up the code of profile.service by removing the unnecessary AuthService.user.subscribe. Implemented the refactoring of some endpoints in HTTP requests which no longer need the user ID in the URL. Default padding and margin in '.row > *' class in home.component.sass now unset for a better alignment. In users-posts.component.sass, fixed the display of images by declaring a new class ".container-slider" with an appropriate setting for margin and overflow. Fixed the reassignment of isLiked$ as an observable in onLikeUp() inside the users-posts.component.ts file. In setting.component, added the new logic for matchPasswordValidator. This commit is part of a task in improving user experience, maintaining clean code and ensuring a better post and likes system for the user interface.