Improve error handling in auth service and login component
The changes have introduced improved error handling in the auth service and the login component. Instead of just throwing the error, we're now throwing the more specific error.message in the case of a failed request. In the login component, errors are now displayed to the user in the form of an alert. Additionally, admin path was added into app-router. In login.html an alert box was added to display error messages. An admin component was also added.