Add profile picture change functionality in settings component
This commit introduces the ability for users to change their profile picture. This is done through a new form input in the settings.component.html which allows users to select a new image file. On submission, the selected file is sent as form data in a POST request to the server for updating the user's profile picture. Also added error handling for registration and navigation to posts after clicking logo in the navbar.