Delete 'admin' component, add 'delete post' feature, update 'show more' condition
The 'admin' module, its test files, and references were removed as the admin functionality will now be handled by a new, more robust system.

Added a "delete post" function accessible by the author of the post. This is done by introducing the deletePost call in profile service and implements it in the profile posts component.

Condition for the "Show More" button in HOME/posts component was updated. It will now show if not on the last page.

Inclusion of unsubscribe steps for newly created subscriptions ensures optimal memory usage.