Add image and video support in post & profile components, CRUD user operations in admin section
This commit includes various updates mainly around improved media handling and the addition of an admin section. The media handling in User-posts, Profile-posts, and main Post component are now updated to support both Images and Videos. This was achieved by having separate ngIf conditions to display media based upon whether it is an image or a video.

The endpoints used for fetching profile images in authService were made more specific and includes fallback options. Both user-posts component and profile-posts component now have proper handling for different media types and ids of images are now fetched more efficiently.

An Admin service and corresponding component were introduced, providing a foundation for managing user activities from an administration perspective. It includes functionalities to get all Users from backend, delete user by id, update user details, and promote user to admin status.

The update mechanism is based on using a form wherein the entries are pre-filled with existing user details and can be updated as required. Also, the admin section now allows for the deletion and promotion of a user to admin. The validation of whether the logged-in user is themselves, is considered before these operations are displayed.